init doc Table
